fix:convert ReusableComponentType to a zod enum
