+++ apps/web/src/components/Editor/hooks/useQueryArgs.ts
@@ -1,34 +1,21 @@
 import { useParams, useSearchParams } from "next/navigation";
 
-// props to chat gpt
-function extractParamValue(
-  pathname: string,
-  path: string,
-  paramName: string
-): string {
-  const regexPattern = pathname.replace(/\[([^\]]+)\]/g, (_match, p1) =>
-    p1 === paramName ? "([\\w-]+)" : "[^/]+"
-  );
-
-  const regex = new RegExp(regexPattern);
-
-  const match = regex.exec(path);
-
-  return match?.[1] ?? "";
-}
-
+// =====================================
+// ⬢ Use String Queries
+// =====================================
+// Extract string value from route or search params
+// Returns the first valid string match or empty string fallback
+// =====================================
 export const useStringQuery = (name: string): string => {
   const params = useParams();
   const searchParams = useSearchParams();
 
-  // First check URL search params (?key=value)
   const searchParam = searchParams?.get(name);
 
   if (searchParam) {
     return searchParam;
   }
 
-  // Then check dynamic route params
   const param = params?.[name];
 
   if (typeof param === "string") {
